+++ engine/include/ProjectileSystem.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
-namespace me::systems {
-	// Handles collisions between Projectiles and Enemies
-	void Projectile_Update(float dt);
+namespace me::projectile {
+
+	void Update(float dt);
+
 }